rename src/features/api/event/types/index.ts -> src/features/api/event/types/dto.ts
